requirements.txt
mysql-replication==0.7
PyMySQL==0.6.7
